initial commit of WARN scraper
Context: The Worker Adjustment and Retraining Notification Act
requires employers to provide advance notice of closings or
layoffs if they affect a certain percentage of employees. Oregon
provides these notifications on a web site. This script is designed
to scrape that website, determine which postings are new, and
then download the applicable information in order to learn about
future closures